Organize file structure for easier readability
