+++ mainAndLists/LearnMoreCard.js
@@ -1,8 +1,8 @@
 import React, { Component } from "react";
 import { Text } from "react-native";
 import { Card, Button as CardButton } from "react-native-elements";
-import NavigationService from "./NavigationService";
-import translate from "./translateService";
+import NavigationService from "../components/NavigationService";
+import translate from "../components/translateService";
 import colors from '../colors';
 
 export default class LearnMoreCard extends Component {
