+++ mainAndLists/QuizCard.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import { Text } from "react-native";
 import { Card, Button as CardButton } from "react-native-elements";
 import NavigationService from "../components/NavigationService";
-import translate from "./translateService";
+import translate from "../components/translateService";
 import colors from '../colors';
 
 export default class QuizCard extends Component {
