+++ mainAndLists/__test__/QuizCard.test.js
@@ -3,10 +3,10 @@ import Adapter from "enzyme-adapter-react-16";
 import React from "react";
 import { Button as CardButton } from "react-native-elements";
 import renderer from "react-test-renderer";
-import NavigationService from "../NavigationService";
+import NavigationService from "../../components/NavigationService";
 import QuizCard from "../QuizCard";
 
-jest.mock("../NavigationService.js");
+jest.mock("../../components/NavigationService.js");
 
 configure({ adapter: new Adapter() });
 describe("QuizCard", () => {
